+++ tests/test.lua
@@ -114,7 +114,7 @@ end
 C_AddOns = {
     GetAddOnMetadata = function(_, field)
         if field == "Version" then
-            return "0.10.0-beta.1"
+            return "0.10.0-beta.2"
         end
     end,
 }